Rename hyperion grid detect then XRC file
